Get rid of config defined ace interact actions, use postInit eh instead
Basically get a small perf boost by only compiling actions when the player gets inside the vehicle. We check the vehicle base type for if it's meant to have actions, then add. 
This should produce working vehicles as long as a 
SelfActions exists on the vehicle, and prevent any memes with assignment.